Refactored the fsScanner into a separate package
Had to introduce a common package to share some types/funcs

The new version does basically the same but is split up in many more functions which I'm told makes things more readable.
It also allows for multiple listeners which will make multi-master easier one day.
It also allows for multiple shares to be implemented in a cleaner way.